Partial revert of nest event simplification
