+++ homeassistant/components/nest/manifest.json
@@ -6,7 +6,7 @@
   "documentation": "https://www.home-assistant.io/integrations/nest",
   "requirements": [
       "python-nest==4.1.0",
-      "google-nest-sdm==0.2.1"
+      "google-nest-sdm==0.2.3"
   ],
   "codeowners": [
       "@awarecan",
