+++ homeassistant/components/nest/sensor_sdm.py
@@ -5,7 +5,7 @@ from typing import Optional
 
 from google_nest_sdm.device import Device
 from google_nest_sdm.device_traits import HumidityTrait, TemperatureTrait
-from google_nest_sdm.event import AsyncEventCallback, EventMessage
+from google_nest_sdm.event import EventMessage
 from google_nest_sdm.exceptions import GoogleNestException
 
 from homeassistant.config_entries import ConfigEntry
@@ -54,7 +54,7 @@ async def async_setup_sdm_entry(
     async_add_entities(entities)
 
 
-class SensorBase(Entity, AsyncEventCallback):
+class SensorBase(Entity):
     """Representation of a dynamically updated Sensor."""
 
     def __init__(self, device: Device):
@@ -80,11 +80,11 @@ class SensorBase(Entity, AsyncEventCallback):
 
     async def async_added_to_hass(self):
         """Run when entity is added to register update signal handler."""
-        self.async_on_remove(self._device.add_event_callback(self))
 
-    async def async_handle_event(self, event_message: EventMessage):
-        """Let Home Assistant know device state has been updated."""
-        self.async_write_ha_state()
+        async def handle_event(event_message: EventMessage):
+            self.async_write_ha_state()
+
+        self.async_on_remove(self._device.add_event_callback(handle_event))
 
 
 class TemperatureSensor(SensorBase):
