+++ requirements_all.txt
@@ -684,7 +684,7 @@ google-cloud-pubsub==2.1.0
 google-cloud-texttospeech==0.4.0
 
 # homeassistant.components.nest
-google-nest-sdm==0.2.1
+google-nest-sdm==0.2.3
 
 # homeassistant.components.google_travel_time
 googlemaps==2.5.1
