+++ requirements_test_all.txt
@@ -355,7 +355,7 @@ google-api-python-client==1.6.4
 google-cloud-pubsub==2.1.0
 
 # homeassistant.components.nest
-google-nest-sdm==0.2.1
+google-nest-sdm==0.2.3
 
 # homeassistant.components.gree
 greeclimate==0.10.3
